+++ CHANGELOG.md
@@ -28,6 +28,8 @@ specification.
   ([#28](https://github.com/mantisbt-plugins/snippets/issues/28))
 - Show announcements without context as disabled on the list page
   ([#30](https://github.com/mantisbt-plugins/snippets/issues/30))
+- Display warning when removing last context
+  ([#31](https://github.com/mantisbt-plugins/snippets/issues/31))
 
 ### Removed